+++ NEWS.md
@@ -7,6 +7,7 @@
     * `reportMD()` is renamed to `reportHTML()`
     *  `filter()` method for `formulas`: `minExplainedFragPeaks` is now called `minExplainedPeaks`
 * Fixed incorrect selection after feature table (or other interactive tables) have been manually re-ordered (reported by Thanh Wang)
+* `groups()` and `as.data.table()` methods for `featureGroups`: optionally consider feature areas instead of peak intensities.
 
 ## May 2019